Using task expiration instead of timeout and making it clear in variable names.
Revocation or expiration mechanism cancels a task if it does not start quickly enough
(e.g., because Celery workers are busy).
Timeout cancels the task after it has started.

The second option makes easier to leave the machine in an inconsistent state
(i.e., with an allocated instance that nobody will ever use).
A soft timeout can be used to capture the timeout exception and treat the
unfinished task, but that is complex to implement, test and mantain.

This commit helps with issue #13.